split Imaging into AbstractImaging and Imaging
